create es helper for owners, #419
